test: refactoring, not done
