--- packages/collector/test/tracing/database/redis/clusterApp.js
@@ -1,171 +1,0 @@
-/*
- * (c) Copyright IBM Corp. 2024
- */
-/* eslint-disable no-console */
-
-'use strict';
-
-// NOTE: c8 bug https://github.com/bcoe/c8/issues/166
-process.on('SIGTERM', () => {
-  process.disconnect();
-  process.exit(0);
-});
-
-require('../../../..')();
-
-const bodyParser = require('body-parser');
-const express = require('express');
-const morgan = require('morgan');
-
-const redis = process.env.REDIS_PKG === 'new' ? require('@redis/client') : require('redis');
-const fetch = require('node-fetch-v2');
-const port = require('../../../test_util/app-port')();
-
-const cls = require('../../../../../core/src/tracing/cls');
-const { isCI } = require('../../../../../core/test/test_util');
-const app = express();
-const logPrefix = `Redis Cluster App (${process.pid}):\t`;
-let connectedToRedis = false;
-const agentPort = process.env.INSTANA_AGENT_PORT;
-
-let nodes = [
-  {
-    url: `redis://${process.env.REDIS_NODE_1}`
-  },
-  {
-    url: `redis://${process.env.REDIS_NODE_2}`
-  },
-  {
-    url: `redis://${process.env.REDIS_NODE_3}`
-  }
-];
-
-let defaults = {
-  url: `redis://${process.env.REDIS_NODE_1}`
-};
-
-// NOTE: we cannot run redis cluster on Tekton https://github.com/bitnami/charts/issues/28894
-if (isCI()) {
-  nodes = [
-    {
-      url: `redis://${process.env.AZURE_REDIS_CLUSTER}`
-    }
-  ];
-
-  defaults = {
-    socket: {
-      tls: true
-    },
-    password: process.env.AZURE_REDIS_CLUSTER_PWD
-  };
-}
-
-// node bin/start-test-containers.js --redis-node-0 --redis-node-1 --redis-node-2
-// docker exec -it 2aaaac7b9112 redis-cli -p 6379 cluster info
-const cluster = redis.createCluster({
-  rootNodes: nodes,
-  useReplicas: false,
-  defaults
-});
-
-cluster.on('error', err => log('Redis Cluster Error', err));
-
-const MAX_TRIES = 50;
-let tries = 0;
-
-const connect = async () => {
-  log(`Connecting to cluster. (${nodes.map(node => node.url).join(', ')})`);
-
-  try {
-    await cluster.connect();
-    log('Connected to cluster');
-    connectedToRedis = true;
-  } catch (err) {
-    log('Failed to connect to cluster', err);
-
-    if (tries >= MAX_TRIES) {
-      log('Max tries reached, exiting.');
-      process.exit(1);
-    }
-
-    log('Retrying...');
-    tries += 1;
-    setTimeout(connect, 1000 * 5);
-  }
-};
-
-connect();
-
-if (process.env.WITH_STDOUT) {
-  app.use(morgan(`${logPrefix}:method :url :status`));
-}
-
-app.use(bodyParser.json());
-
-app.get('/', (req, res) => {
-  if (!connectedToRedis) {
-    res.sendStatus(500);
-  } else {
-    res.sendStatus(200);
-  }
-});
-
-app.post('/clearkeys', async (req, res) => {
-  if (cls.isTracing()) {
-    cls.setTracingLevel('0');
-  }
-
-  await Promise.all(
-    cluster.masters.map(async master => {
-      const client = await cluster.nodeClient(master);
-      await client.flushAll();
-    })
-  );
-
-  if (cls.isTracing()) {
-    cls.setTracingLevel('1');
-  }
-
-  res.sendStatus(200);
-});
-
-app.post('/values', async (req, res) => {
-  const key = req.query.key;
-  const value = req.query.value;
-
-  try {
-    await cluster.set(key, value);
-    log('Set key successfully.');
-  } catch (e) {
-    log('Set with key %s, value %s failed', key, value, e);
-    res.sendStatus(500);
-  }
-  await fetch(`http://127.0.0.1:${agentPort}`);
-  log('Sent agent fetch successfully.');
-  res.sendStatus(200);
-});
-
-app.get('/values', async (req, res) => {
-  const key = req.query.key;
-
-  try {
-    const redisRes = await cluster.get(key);
-    log('Got redis key successfully.');
-    await fetch(`http://127.0.0.1:${agentPort}`);
-    log('Sent agent request successfully.');
-    res.send(redisRes);
-  } catch (err) {
-    log('Get with key %s failed', key, err);
-    res.sendStatus(500);
-  }
-});
-
-app.listen(port, () => {
-  log(`Listening on port: ${port}`);
-});
-
-function log() {
-  const args = Array.prototype.slice.call(arguments);
-  args[0] = logPrefix + args[0];
-  console.log.apply(console, args);
-}
