+++ packages/collector/test/tracing/database/redis/connect-via/index.js
@@ -1,0 +1,13 @@
+/*
+ * (c) Copyright IBM Corp. 2024
+ */
+
+'use strict';
+
+module.exports = function (redis, log) {
+  if (process.env.REDIS_CLUSTER === 'true') {
+    return require('./cluster')(redis, log);
+  }
+
+  return require('./default')(redis, log);
+};
