+++ packages/collector/test/tracing/database/redis/legacyApp.js
@@ -13,6 +13,8 @@ process.on('SIGTERM', () => {
   process.exit(0);
 });
 
+require('./mockVersion');
+
 require('../../../..')();
 
 const bodyParser = require('body-parser');
